requirements.txt
Flask==2.3.2
cryptography==41.0.3
PyNaCl==1.5.0